+++ examples/resources/konnect_auth_server/resource.tf
@@ -1,13 +1,14 @@
 resource "konnect_auth_server" "my_authserver" {
   provider = konnect-beta
-  audience      = "...my_audience..."
-  description   = "...my_description..."
-  force_destroy = "true"
+  audience                          = "...my_audience..."
+  dcr_default_access_token_duration = 300
+  description                       = "...my_description..."
+  force_destroy                     = "false"
   labels = {
     key = "value"
   }
   name              = "...my_name..."
-  signing_algorithm = "RS384"
+  signing_algorithm = "RS256"
   trusted_origins = [
     "https://example.com"
   ]