+++ examples/resources/konnect_event_gateway_vault/resource.tf
@@ -4,7 +4,7 @@ resource "konnect_event_gateway_vault" "my_eventgatewayvault" {
     config = {
       prefix = "KONG_"
     }
-    description = "...my_description..."
+    description = ""
     labels = {
       key = "value"
     }
@@ -12,7 +12,7 @@ resource "konnect_event_gateway_vault" "my_eventgatewayvault" {
   }
   gateway_id = "9524ec7d-36d9-465d-a8c5-83a3c9390458"
   konnect = {
-    description = "...my_description..."
+    description = ""
     labels = {
       key = "value"
     }