+++ examples/resources/konnect_mesh_external_service/resource.tf
@@ -24,7 +24,7 @@ resource "konnect_mesh_external_service" "my_meshexternalservice" {
     }
     tls = {
       allow_renegotiation = false
-      enabled             = true
+      enabled             = false
       verification = {
         ca_cert = {
           inline        = "...my_inline..."
@@ -41,7 +41,7 @@ resource "konnect_mesh_external_service" "my_meshexternalservice" {
           inline_string = "...my_inline_string..."
           secret        = "...my_secret..."
         }
-        mode        = "SkipAll"
+        mode        = "Secured"
         server_name = "...my_server_name..."
         subject_alt_names = [
           {
@@ -51,8 +51,8 @@ resource "konnect_mesh_external_service" "my_meshexternalservice" {
         ]
       }
       version = {
-        max = "TLS12"
-        min = "TLS10"
+        max = "TLSAuto"
+        min = "TLSAuto"
       }
     }
   }