+++ internal/sdk/internal/hooks/registration.go
@@ -8,6 +8,13 @@ func initHooks(h *Hooks) {
 	h.registerBeforeRequestHook(&MeshDefaultsHook{})
 	h.registerBeforeRequestHook(&AuthServerClientCreateHook{})
 
+	// Domain customization - enable usage with non-prod domains
+	h.registerBeforeRequestHook(&CustomizeKongDomainHook{
+		Enabled:           os.Getenv("KONG_CUSTOM_DOMAIN") != "",
+		Domain:            os.Getenv("KONG_CUSTOM_DOMAIN"),
+		ReplaceFullDomain: os.Getenv("KONG_CUSTOM_DOMAIN_REPLACE_FULL_DOMAIN") == "true",
+	})
+
 	// Debug hooks - dump request/response
 	h.registerBeforeRequestHook(&HTTPDumpRequestHook{
 		Enabled: os.Getenv("KONNECT_SDK_HTTP_DUMP_REQUEST") == "true",