+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 
 ## 0.16.0
-> Released on 2026/02/??
+> Released on 2026/02/27
 
 ### BREAKING CHANGES
 * `event_gateway_listener_id` renamed to `listener_id` in `konnect_event_gateway_listener_policy_forward_to_virtual_cluster` and `konnect_event_gateway_listener_policy_tls_server`
 * Schema registry references simplified to use `id` directly instead of `schema_registry_reference_by_id`/`schema_registry_reference_by_name` in `konnect_event_gateway_consume_policy_schema_validation` and `konnect_event_gateway_produce_policy_schema_validation`
 * Virtual cluster references simplified to use `id` directly instead of `virtual_cluster_reference_by_id`/`virtual_cluster_reference_by_name` in `konnect_event_gateway_listener_policy_forward_to_virtual_cluster`
 * Static key references simplified to use `id` directly instead of `reference_by_id`/`reference_by_name` in `konnect_event_gateway_produce_policy_encrypt`
+
+### Features
+* Add support for configuring Managed Cache in Cloud Gateway using `konnect_cloud_gateway_add_on` resource
 
 ## 0.15.0
 > Released on 2026/02/16